+++ pgtap/tsp/TSPeuclidean/types_check.sql
@@ -4,8 +4,8 @@
 SELECT plan(5);
 SET client_min_messages TO WARNING;
 
-SELECT has_function('pgr_tspeuclidean');
-SELECT has_function('pgr_tspeuclidean', ARRAY[
+SELECT has_function('pgr_tspannealingeuclidean');
+SELECT has_function('pgr_tspannealingeuclidean', ARRAY[
     'text', 'bigint', 'bigint',
     'double precision',
     'integer', 'integer', 'integer',
@@ -14,7 +14,7 @@ SELECT has_function('pgr_tspeuclidean', ARRAY[
     'double precision',
     'boolean'
     ]);
-SELECT function_returns('pgr_tspeuclidean', ARRAY[
+SELECT function_returns('pgr_tspannealingeuclidean', ARRAY[
     'text', 'bigint', 'bigint',
     'double precision',
     'integer', 'integer', 'integer',
@@ -26,7 +26,7 @@ SELECT function_returns('pgr_tspeuclidean', ARRAY[
 
 -- parameter names
 SELECT set_eq(
-    $$SELECT proargnames FROM pg_proc WHERE proname = 'pgr_tspeuclidean'$$,
+    $$SELECT proargnames FROM pg_proc WHERE proname = 'pgr_tspannealingeuclidean'$$,
     $$SELECT '{
         "","start_id","end_id",
         "max_processing_time",
@@ -38,7 +38,7 @@ SELECT set_eq(
 
 -- parameter types
 SELECT set_eq(
-    $$SELECT proallargtypes FROM pg_proc WHERE proname = 'pgr_tspeuclidean'$$,
+    $$SELECT proallargtypes FROM pg_proc WHERE proname = 'pgr_tspannealingeuclidean'$$,
     $$SELECT '{25,20,20,701,23,23,23,701,701,701,16,23,20,701,701}'::OID[] $$
 );