+++ sql/tsp/TSPannealing.sql
@@ -24,9 +24,14 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 ********************************************************************PGR-GNU*/
 
+-------------
+-- pgr_TSP
+-------------
 
-CREATE OR REPLACE FUNCTION _pgr_TSPeuclidean(
-    coordinates_sql TEXT,
+
+CREATE OR REPLACE FUNCTION pgr_TSPannealing(
+    matrix_row_sql TEXT,
+
     start_id BIGINT DEFAULT 0,
     end_id BIGINT DEFAULT 0,
 
@@ -42,15 +47,39 @@ CREATE OR REPLACE FUNCTION _pgr_TSPeuclidean(
 
     randomize BOOLEAN DEFAULT true,
 
-    OUT seq integer,
+    OUT seq INTEGER,
     OUT node BIGINT,
     OUT cost FLOAT,
     OUT agg_cost FLOAT)
-RETURNS SETOF record
-AS 'MODULE_PATHNAME', 'euclideanTSP'
-LANGUAGE C VOLATILE STRICT;
+RETURNS SETOF RECORD AS
+$BODY$
+    SELECT *
+    FROM _pgr_TSPannealing(_pgr_get_statement($1), $2, $3, $4, $5, $6, $7, $8, $9, $10, $11);
+$BODY$
+LANGUAGE SQL VOLATILE STRICT
+COST 100
+ROWS 1000;
+
 
 -- COMMENTS
 
-COMMENT ON FUNCTION _pgr_TSPeuclidean(TEXT, BIGINT, BIGINT, FLOAT, INTEGER, INTEGER, INTEGER, FLOAT, FLOAT, FLOAT, BOOLEAN)
-IS 'pgRouting internal function';
+
+COMMENT ON FUNCTION pgr_TSPannealing(TEXT, BIGINT, BIGINT, FLOAT, INTEGER, INTEGER, INTEGER, FLOAT, FLOAT, FLOAT, BOOLEAN)
+IS 'pgr_TSP
+ - Parameters
+   - matrix SQL with columns: start_vid, end_vid, agg_cost
+ - Optional parameters
+    - start_id := 0
+    - end_id := 0
+
+    - max_processing_time := ''+infinity''::FLOAT
+
+    - tries_per_temperature := 500
+    - max_changes_per_temperature :=  60
+    - max_consecutive_non_changes :=  100
+
+    - initial_temperature FLOAT := 100
+    - final_temperature := 0.1
+    - cooling_factor := 0.9
+
+    - randomize := true';