+++ test/tsp/doc-pgr_TSPannealingEuclidean.test.sql
@@ -1,11 +1,11 @@
 \echo -- q1
-SELECT * FROM pgr_TSPeuclidean(
+SELECT * FROM pgr_TSPannealingEuclidean(
     $$
     SELECT id, st_X(the_geom) AS x, st_Y(the_geom)AS y  FROM edge_table_vertices_pgr
     $$,
     randomize := false);
 \echo -- q2
-SELECT* from pgr_TSPeuclidean(
+SELECT* from pgr_TSPannealingEuclidean(
     $$
     SELECT id, st_X(the_geom) AS x, st_Y(the_geom) AS y FROM edge_table_vertices_pgr
     $$,
@@ -15,7 +15,7 @@ SELECT* from pgr_TSPeuclidean(
 \echo -- q3
 SET client_min_messages TO DEBUG1;
 
-SELECT* from pgr_TSPeuclidean(
+SELECT* from pgr_TSPannealingEuclidean(
     $$
     SELECT id, st_X(the_geom) AS x, st_Y(the_geom) AS y FROM edge_table_vertices_pgr
     $$,