+++ doc/src/release_notes.rst
@@ -53,7 +53,9 @@ pgRouting 3.0.0 Release Notes
 
 .. rubric:: Fixed Issues
 
-* `#1153 <https://github.com/pgRouting/pgrouting/issues/1153>`__: Renamed pgr_eucledianTSP to pgr_TSPeuclidean
+* `#1153 <https://github.com/pgRouting/pgrouting/issues/1153>`__:
+  * Renamed pgr_eucledianTSP to pgr_TSPannealingEuclidean
+  * Renamed pgr_TSP to pgr_TSPannealing
 
 .. rubric:: New Experimental functions